--- src/lib/schemas/lists.json
@@ -1,81 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "$id": "https://cra.orcwg.org/schemas/lists.json",
-  "title": "CRA Curated List Item",
-  "description": "Schema for processed curated list items in the CRA system",
-  "type": "object",
-  "properties": {
-    "answer": {
-      "type": ["string", "null"],
-      "description": "Parsed answer/description from the list file"
-    },
-    "category": {
-      "type": "string",
-      "description": "The category/directory this list belongs to (usually 'root' for lists)"
-    },
-    "content": {
-      "type": "string",
-      "description": "Raw markdown content of the list file"
-    },
-    "description": {
-      "type": "string",
-      "description": "Brief description of what this list covers"
-    },
-    "items": {
-      "type": "array",
-      "description": "Array of FAQ items included in this curated list",
-      "minItems": 1,
-      "items": {
-        "type": "object",
-        "properties": {
-          "question": {
-            "type": "string",
-            "description": "The question text from the FAQ"
-          },
-          "url": {
-            "type": "string",
-            "description": "The URL path to the FAQ item",
-            "pattern": "^/faq/"
-          }
-        },
-        "required": ["question", "url"],
-        "additionalProperties": false
-      }
-    },
-    "order": {
-      "type": "integer",
-      "description": "Display order (lower numbers appear first)",
-      "minimum": 1
-    },
-    "question": {
-      "type": ["string", "null"],
-      "description": "Parsed question/title from the list file"
-    },
-    "title": {
-      "type": "string",
-      "description": "Display name for the curated list"
-    }
-  },
-  "required": ["title", "items"],
-  "additionalProperties": true,
-  "examples": [
-    {
-      "title": "Getting Started with CRA",
-      "description": "Essential FAQs for understanding the basics of the Cyber Resilience Act",
-      "order": 1,
-      "items": [
-        {
-          "question": "What is the Cyber Resilience Act (CRA)?",
-          "url": "/faq/cra-itself/cra/"
-        },
-        {
-          "question": "What is the scope of the CRA?",
-          "url": "/faq/cra-itself/scope/"
-        }
-      ],
-      "content": "---\ntitle: \"Getting Started with CRA\"\norder: 1\nfaqs:\n\n- cra-itself/cra\n- cra-itself/scope\n\n---\n\nEssential FAQs for understanding the basics of the Cyber Resilience Act",
-      "question": "Getting Started with CRA",
-      "answer": "Essential FAQs for understanding the basics of the Cyber Resilience Act"
-    }
-  ]
-}
