+++ src/lists.njk
@@ -10,10 +10,10 @@ permalink: "/lists/"
 </div>
 
 <div class="curated-lists-grid">
-  {% for listItem in orcwgImport.lists %}
+  {% for listItem in orcwgImport.list %}
     <a href="{{ listItem.url }}" class="curated-list-card">
       <h2 class="list-title">{{ listItem.title }}</h2>
-      <p class="list-description">{{ listItem.templateData.description }}</p>
+      <p class="list-description">{{ listItem.description }}</p>
     </a>
   {% endfor %}
 </div>
